+++ src/Drone/Types/Log.hs
@@ -4,10 +4,11 @@
 module Drone.Types.Log where
 
 import           Data.Extensible
+import           Data.Int        (Int64)
 import           Data.Text       (Text)
 
 type Line = Record
    '[ "pos"  >: Int
     , "out"  >: Text
-    , "time" >: Int
+    , "time" >: Int64
     ]
